requirements.txt
psycopg2-binary==2.9.9
faker==24.3.0
python-dotenv==1.0.1
tenacity==8.2.3
